# Conflicts:
#	src/main/java/com/keymanager/ckadmin/controller/internal/rest/QZSettingController.java
#	src/main/webapp/WEB-INF/views/qzsettings/qzsetting.html

+++ src/main/java/com/keymanager/ckadmin/service/impl/CustomerServiceImpl.java
@@ -170,7 +170,7 @@ public class CustomerServiceImpl extends ServiceImpl<CustomerDao, Customer> impl
     }
 
     @Override
-    public Customer getCustomer(int customerUuid) {
+    public Customer getCustomer(Long customerUuid) {
         return customerDao.selectById(customerUuid);
     }
 }
